Adds offers list component
